Develop BettingOdds toDTO Mapper method

+++ src/main/java/com/haveanicepickem/restservice/dto/BettingOddResponseDTO.java
@@ -4,7 +4,7 @@ package com.haveanicepickem.restservice.dto;
 public record BettingOddResponseDTO(
     String source,
     Float overUnder,
-    short moneyline,
+    byte moneyline,
     Float spread,
     Float winProbability
 ) {}
